Implemented adjustEditTextAnswerToStringNumberWidget() method
